#27 feat. 메인 페이지 탭 레이아웃 추가
