+++ app/src/main/res/layout/fragment_home_articles_list.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.core.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/background_default_secondary">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/articles_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="30dp"
+        android:paddingBottom="30dp" />
+
+</androidx.core.widget.NestedScrollView>
